remove the need for idx in extension models
the twich api has a very weird/annoying way of configuring the json request for this (or maybe I just dont fully understand it, which is also pretty likely). this method is a bit cleaner as it automatically fills the 'index' key. maybe still need some work as i think the extension models dont have to be fully defined when updated. ill re-visit this again when im making the twitch client wrapper around this API